backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
djangorestframework==3.15.2
django-cors-headers
djangorestframework-simplejwt==5.5.0

